Added Layout to Project
Some things to note:
Need to look at C3 and change it to a polar cap if necessary.
Should we add more than one 3v3 power rail?
EEPROM closer to chip?

TODO:
check supporting circuitry for ICs (namely power and CAN and EEPROM) to match manufacturers
recommended implementation.
Submit to be outsourced for manufacturing.

Laid out the schematic so the board is 3.5in x 3.5in (89mmx89mm).
Added GPIO pinout headers to make use easier.
Standoff hole sizing is 3.5mm in diameter (M3 Stand-off ready)

BOM will be submitted in next for commits for ordering.

Only a B. Ground plane exists. F. Cu 3V3 rail was considered but decided against
due to islands isolating large parts of circuitry that needed to be connected.

Still needs to be reviewed by David, and okayed with Keona and Mike in terms of layout
and access to power.