Added fix for power OVLD channels and added power flags.
The proper design for a comparator with hysteresis has pull-up
resistors on the output of the comparator. However these were forgotten
on the HELIX_Power_Prototype board, so they can be added here for testing.
Test points were also added to make it easier to debug the comparator.
Power flags were added to remove some of the DRC errors, the remaining
errors require changes to the SAME51J20A symbol.