Fixed power issue on SAME51J20A
The VDDCORE pin is actually a voltage output of ~1.2v that is
intended to supply power to the VSW pin when in Switching mode.
The SAME51J20A symbol was fixed in the helix symbol library to
indicate the VDDCORE pin as a voltage output rather than an input
and the schematic was also changed to reflect the recommended
power supply setup from the SAME5x datasheet.